+++ Ray-Tracing/geometry.cpp
@@ -11,10 +11,10 @@ Vector Vector::operator-() {
     return Vector(-x_, -y_, -z_);
 }
 
-long double Vector::length() {
+double Vector::length() {
     return std::sqrt(sqrLength());
 }
-long double Vector::sqrLength() {
+double Vector::sqrLength() {
     return x_ * x_ + y_ * y_ + z_ * z_;
 }
 Vector Vector::normed() {
@@ -23,16 +23,16 @@ Vector Vector::normed() {
 Vector Vector::crossProduct(Vector v) {
     return Vector(y_ * v.z_ - v.y_ * z_, v.x_ * z_ - x_ * v.z_, x_ * v.y_ - v.x_ * y_);
 }
-long double Vector::dotProduct(Vector v) {
+double Vector::dotProduct(Vector v) {
     return v.x_ * x_ + v.y_ * y_ + v.z_ * z_;
 }
-long double dotProduct(Vector a, Vector b) {
+double dotProduct(Vector a, Vector b) {
     return a.x_ * b.x_ + a.y_ * b.y_ + a.z_ * b.z_;
 }
-long double Vector::orientVolume(Vector a, Vector b) {
+double Vector::orientVolume(Vector a, Vector b) {
     return x_ * (a.y_ * b.z_ - a.z_ * b.y_) - y_ * (a.x_ * b.z_ - a.z_ * b.x_) + z_ * (a.x_ * b.y_ - a.y_ * b.x_);
 }
-long double Vector::absVolume(Vector a, Vector b) {
+double Vector::absVolume(Vector a, Vector b) {
     return std::abs(orientVolume(a, b));
 }
 int Vector::checkColinear(Vector a) {
@@ -42,7 +42,7 @@ int Vector::checkColinear(Vector a) {
         return -1;
     return 0;
 }
-long double absVolume(Vector a, Vector b, Vector c) {
+double absVolume(Vector a, Vector b, Vector c) {
     return a.absVolume(b, c);
 }
 int checkColinear(Vector a, Vector b) {
@@ -61,13 +61,13 @@ Vector operator- (Vector left, const Vector& right) {
     return tmp;
 }
 
-Vector operator* (Vector left, const long double c) {
+Vector operator* (Vector left, const double c) {
     Vector tmp(left);
     tmp.x_ *= c; tmp.y_ *= c; tmp.z_ *= c;
     return tmp;
 }
 
-Vector operator*(const long double c, Vector left) {
+Vector operator*(const double c, Vector left) {
     Vector tmp(left);
     tmp.x_ *= c; tmp.y_ *= c; tmp.z_ *= c;
     return tmp;
@@ -81,11 +81,11 @@ bool operator!=(const Vector& lhs, const Vector& rhs) {
     return !operator==(lhs, rhs);
 }
 
-Vector operator/(Vector left, const long double c) {
+Vector operator/(Vector left, const double c) {
     return left * (1.0 / c);
 }
 
-Vector operator/(const long double c, Vector rhs) {
+Vector operator/(const double c, Vector rhs) {
     return (1.0 / c) * rhs;
 }
